chore(layout): split oversized prod files into submodules [#35]
Three targeted splits within the 700L/1000L gates:

1. khive-pack-knowledge: wire orphaned scoring.rs + de-duplicate
   search.rs extracted Weights/Candidate/scoring fns from search.rs
   (was duplicated inline). Wire existing scoring.rs via mod.rs.
   search.rs: 1661L → 1418L (-243L); scoring.rs: was orphaned, now wired.

2. khive-pack-kg: extract param structs from common.rs into params.rs
   All deserialization structs + tri_f64 moved to handlers/params.rs (231L).
   common.rs: 1025L → 806L (-219L). Re-exports preserve all consumers.

3. khive-runtime: extract config types from runtime.rs into config.rs
   BackendId, NamespaceToken, RuntimeConfig, and all embedding helpers
   moved to runtime/config.rs (382L).
   runtime.rs: 1333L → 971L (-362L). pub use re-exports maintain crate API.

Also fix: GateRef missing from runtime tests (import moved to test scope);
fix: retrieval.rs imports sanitize_key/parse_embedding_model_alias from
crate::config not crate::runtime; remove dead top_k_sorted fn.

Verification: cargo check, clippy -D warnings, test --workspace, fmt --check
all pass with 0 errors 0 warnings.
