+++ crates/khive-pack-knowledge/src/knowledge/mod.rs
@@ -2,6 +2,7 @@
 
 pub(crate) mod matching;
 pub(crate) mod schema;
+pub(crate) mod scoring;
 pub(crate) mod section_feedback;
 pub(crate) mod vamana;
